[MTSDK-1261] move TLS helper functions to separate file to fix duplicate PlatformSocketKt facade
